chore: rename contracts and related files
