+++ contracts/oracles/simple-oracle.clar
