+++ tests/usr-entry.test.ts
